Enhance Odie notice visibility handling and refactor components for improved user experience
