+++ packages/odie-client/src/components/message/messages-container.tsx
@@ -1,22 +1,18 @@
 import { useResetSupportInteraction } from '@automattic/help-center/src/hooks/use-reset-support-interaction';
-import { HELP_CENTER_STORE } from '@automattic/help-center/src/stores';
 import { getShortDateString } from '@automattic/i18n-utils';
 import { Spinner } from '@wordpress/components';
-import { useDispatch as useDataStoreDispatch } from '@wordpress/data';
-import { __, _n } from '@wordpress/i18n';
+import clsx from 'clsx';
 import { useEffect, useRef, useState } from 'react';
-import { useLocation, useNavigate, useSearchParams } from 'react-router-dom';
+import { useSearchParams } from 'react-router-dom';
 import { ThumbsDown } from '../../assets/thumbs-down';
 import { useOdieAssistantContext } from '../../context';
-import { useGetSupportInteractionById } from '../../data/use-get-support-interaction-by-id';
 import {
 	useAutoScroll,
 	useCreateZendeskConversation,
 	useZendeskMessageListener,
 } from '../../hooks';
-import { useGetMostRecentOpenConversation } from '../../hooks/use-get-most-recent-open-conversation';
 import { getOdieInitialMessage } from '../../utils';
-import OdieNotice from '../odie-notice';
+import { ViewMostRecentOpenConversationNotice } from '../odie-notice/view-most-recent-conversation-notice';
 import { DislikeFeedbackMessage } from './dislike-feedback-message';
 import { JumpToRecent } from './jump-to-recent';
 import { ThinkingPlaceholder } from './thinking-placeholder';
@@ -47,58 +43,6 @@ const ChatDate = ( { chat }: { chat: Chat } ) => {
 	return <div className="odie-chat__date">{ currentDate }</div>;
 };
 
-const ViewMostRecentOpenConversationNotice = () => {
-	const { mostRecentSupportInteractionId, totalNumberOfConversations } =
-		useGetMostRecentOpenConversation();
-
-	const fetchSupportInteraction =
-		mostRecentSupportInteractionId?.toString() && totalNumberOfConversations === 1
-			? mostRecentSupportInteractionId.toString()
-			: null;
-	const { data: supportInteraction } = useGetSupportInteractionById( fetchSupportInteraction );
-	const { setCurrentSupportInteraction } = useDataStoreDispatch( HELP_CENTER_STORE );
-	const { trackEvent } = useOdieAssistantContext();
-	const location = useLocation();
-	const navigate = useNavigate();
-	const shouldDisplayNotice = supportInteraction || totalNumberOfConversations > 1;
-
-	const handleNoticeOnClick = () => {
-		if ( supportInteraction ) {
-			setCurrentSupportInteraction( supportInteraction );
-			if ( ! location.pathname.includes( '/odie' ) ) {
-				navigate( '/odie' );
-			}
-		} else {
-			navigate( '/chat-history' );
-		}
-		trackEvent( 'chat_open_previous_conversation_notice', {
-			destination: supportInteraction ? 'support-interaction' : 'chat-history',
-			total_number_of_conversations: totalNumberOfConversations,
-		} );
-	};
-
-	return (
-		shouldDisplayNotice && (
-			<OdieNotice>
-				<div className="odie-notice__view-conversation">
-					<span>
-						{ __( 'You have another open conversation already started.', __i18n_text_domain__ ) }
-					</span>
-					&nbsp;
-					<button onClick={ handleNoticeOnClick }>
-						{ _n(
-							'View conversation',
-							'View conversations',
-							totalNumberOfConversations,
-							__i18n_text_domain__
-						) }
-					</button>
-				</div>
-			</OdieNotice>
-		)
-	);
-};
-
 interface ChatMessagesProps {
 	currentUser: CurrentUser;
 }
@@ -113,6 +57,10 @@ export const MessagesContainer = ( { currentUser }: ChatMessagesProps ) => {
 	const [ hasForwardedToZendesk, setHasForwardedToZendesk ] = useState( false );
 	const [ chatMessagesLoaded, setChatMessagesLoaded ] = useState( false );
 	const messagesContainerRef = useRef< HTMLDivElement >( null );
+	const viewConversationNoticeRef = useRef< HTMLDivElement >( null );
+
+	const [ isNoticeVisible, setIsNoticeVisible ] = useState( false );
+
 	useZendeskMessageListener();
 	useAutoScroll( messagesContainerRef );
 
@@ -158,6 +106,36 @@ export const MessagesContainer = ( { currentUser }: ChatMessagesProps ) => {
 		createZendeskConversation,
 	] );
 
+	// Detect odie notice visibility
+	useEffect( () => {
+		const observer = new MutationObserver( ( mutations ) => {
+			mutations.forEach( ( mutation ) => {
+				if ( mutation.type === 'childList' ) {
+					const noticeElement = document.querySelector( '.odie-notice' );
+					if ( noticeElement ) {
+						const isVisible =
+							noticeElement.getBoundingClientRect().top < window.innerHeight &&
+							noticeElement.getBoundingClientRect().bottom >= 0;
+						setIsNoticeVisible( isVisible );
+					} else {
+						setIsNoticeVisible( false );
+					}
+				}
+			} );
+		} );
+
+		const targetNode = document.querySelector( '.chatbox-messages' );
+		if ( targetNode ) {
+			observer.observe( targetNode, { childList: true, subtree: true } );
+		}
+
+		return () => {
+			if ( targetNode ) {
+				observer.disconnect();
+			}
+		};
+	}, [] );
+
 	// Used to apply the correct styling on messages
 	const isNextMessageFromSameSender = ( currentMessage: string, nextMessage: string ) => {
 		return currentMessage === nextMessage;
@@ -171,7 +149,12 @@ export const MessagesContainer = ( { currentUser }: ChatMessagesProps ) => {
 
 	return (
 		<>
-			<div className="chatbox-messages" ref={ messagesContainerRef }>
+			<div
+				className={ clsx( 'chatbox-messages', {
+					'is-notice-visible': isNoticeVisible,
+				} ) }
+				ref={ messagesContainerRef }
+			>
 				<ChatDate chat={ chat } />
 				{ ! chatMessagesLoaded ? (
 					<LoadingChatSpinner />
@@ -199,7 +182,9 @@ export const MessagesContainer = ( { currentUser }: ChatMessagesProps ) => {
 							/>
 						) ) }
 						<JumpToRecent containerReference={ messagesContainerRef } />
-						{ chat.provider === 'odie' && <ViewMostRecentOpenConversationNotice /> }
+						{ chat.provider === 'odie' && (
+							<ViewMostRecentOpenConversationNotice ref={ viewConversationNoticeRef } />
+						) }
 						{ chat.status === 'dislike' && ! removeDislikeStatus && <DislikeThumb /> }
 						{ availableStatusWithFeedback.includes( chat.status ) && (
 							<div className="odie-chatbox__action-message">
